Vendor Dependencies
- This creates git submodules for the project dependencies in `vendor/`
- During build we tick the `GO15VENDOREXPERIMENT` env var, ang Go knows to look
in `vendor/` for the dependencies, rather than on the `$GOPATH`.
- Something similar can be accomplished with tools like `godep`, but they rewrite your
import paths, and also didn't work for me :D
- This is cool and fine, except that when you clone now, you either need to:
`git clone --recursive` or `make dependencies` in the repo.
- Overall I think it's a good trade-off, and I can delete `rackerlabs/dns` because that's
just sillypants.

- This patch also includes some changes to the Dockerfile, they recommend starting with the
`onbuild` one, which we had. But it makes certain assumptions about your code that weren't
being met anymore, especially with these deps.
- Other changes for the dockerfile: make sure it builds the current branch, and use `make build`, which
injects the metadata for `--version` to work right.